requirements.txt
mysql-connector
pyyaml
configparser
telepot
